requirements.txt
tensorflow==2.12.0
pandas==2.1.3
dvc==3.47.0
mlflow==2.2.2
numpy==1.23.5
matplotlib==3.8.2
python-box==6.0.2
pyYAML==6.0.1
tqdm==4.66.1
ensure==1.0.2
joblib==1.3.2
types-PyYAML==6.0.12.12
scipy==1.11.4
Flask==2.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
Werkzeug==3.0.1